+++ mom6_forge/grid.py
@@ -9,7 +9,7 @@ from mom6_forge._supergrid import (
     RectilinearCartesianSupergrid,
     SupergridBase,
 )
-from mom6_forge.utils import normalize_deg
+from mom6_forge.utils import normalize_deg, mom6_angle_calculation_method
 
 
 class Grid:
@@ -518,6 +518,47 @@ class Grid:
         obj._compute_MOM6_grid_metrics()
 
         return obj
+
+    def calculate_t_point_rotation_angles_from_mom6_method(self) -> xr.DataArray:
+        """
+        Calculate the ``angle_dx`` in degrees from the true ``x`` direction (parallel to latitude) counter-clockwise
+        and return as a dataarray. (Mimics MOM6 angle calculation function :func:`~mom6_angle_calculation_method`)
+
+        Use this to compare with the angle_dx provided in the supergrid file to check if the angles are consistent with MOM6 method
+
+        Parameters
+        ----------
+        supergrid: xr.Dataset
+            The supergrid dataset
+
+        Returns
+        -------
+        xr.DataArray
+            The t-point angles
+        """
+        # t-points: cell centers at every other supergrid point starting at index 1
+        t_points = xr.Dataset(
+            {
+                "x": (("nyp", "nxp"), self._supergrid.x.values[1::2, 1::2]),
+                "y": (("nyp", "nxp"), self._supergrid.y.values[1::2, 1::2]),
+            }
+        )
+        # q-points: cell corners at every other supergrid point starting at index 0
+        q_points = xr.Dataset(
+            {
+                "x": (("nyp", "nxp"), self._supergrid.x.values[0::2, 0::2]),
+                "y": (("nyp", "nxp"), self._supergrid.y.values[0::2, 0::2]),
+            }
+        )
+
+        return mom6_angle_calculation_method(
+            self._supergrid.x.max() - self._supergrid.x.min(),
+            q_points.isel(nyp=slice(1, None), nxp=slice(0, -1)),
+            q_points.isel(nyp=slice(1, None), nxp=slice(1, None)),
+            q_points.isel(nyp=slice(0, -1), nxp=slice(0, -1)),
+            q_points.isel(nyp=slice(0, -1), nxp=slice(1, None)),
+            t_points,
+        )
 
     @classmethod
     def subgrid_from_supergrid(
