+++ mom6_forge/utils.py
@@ -697,3 +697,117 @@ def quadrilateral_area(v1, v2, v3, v4):
     a4 = angle_between(v4, v1, v3)
 
     return (a1 + a2 + a3 + a4 - 2 * np.pi) * R**2
+
+
+def mom6_angle_calculation_method(
+    len_lon,
+    top_left: xr.DataArray,
+    top_right: xr.DataArray,
+    bottom_left: xr.DataArray,
+    bottom_right: xr.DataArray,
+    point: xr.DataArray,
+) -> xr.DataArray:
+    """
+    Calculate the angle of the grid point's local x-direction compared to East-West direction
+    using the MOM6 method adapted from: https://github.com/mom-ocean/MOM6/blob/05d8cc395c1c3c04dd04885bf8dd6df50a86b862/src/initialization/MOM_shared_initialization.F90#L572-L587
+
+    Note: this is exactly the same as the angle of the grid point's local y-direction compared to North-South direction.
+
+    This method can handle vectorized computations.
+
+    Parameters
+    ----------
+    len_lon: float
+        The extent of the longitude of the regional domain (in degrees).
+    top_left, top_right, bottom_left, bottom_right: xr.DataArray
+        The four points around the point to calculate the angle from the ``supergrid``;
+        requires both an ``x``` and ``y`` component (both in degrees).
+    point: xr.DataArray
+        The point to calculate the angle from the ``supergrid``
+
+    Returns
+    -------
+    xr.DataArray
+        The angle of the grid point's local ``x``-direction compared to East-West direction.
+    """
+
+    # Compute lonB for all points
+    lonB = np.zeros((2, 2, len(point.nyp), len(point.nxp)))
+
+    # Vectorized computation of lonB
+    lonB[0][0] = modulo_around_point(bottom_left.x, point.x, len_lon)  # Bottom Left
+    lonB[1][0] = modulo_around_point(top_left.x, point.x, len_lon)  # Top Left
+    lonB[1][1] = modulo_around_point(top_right.x, point.x, len_lon)  # Top Right
+    lonB[0][1] = modulo_around_point(bottom_right.x, point.x, len_lon)  # Bottom Right
+
+    cos_meanlat = np.cos(
+        np.deg2rad((bottom_left.y + bottom_right.y + top_right.y + top_left.y) / 4)
+    )
+
+    # Quadrilateral diagonals
+
+    # top-left--bottom-right diagonal components
+    TL_BR_diagonal_x = cos_meanlat * (lonB[1, 0] - lonB[0, 1])
+    TL_BR_diagonal_y = top_left.y - bottom_right.y
+
+    # top-right--bottom-left diagonal components
+    TR_BL_diagonal_x = cos_meanlat * (lonB[1, 1] - lonB[0, 0])
+    TR_BL_diagonal_y = top_right.y - bottom_left.y
+
+    # Sum of diagonals components
+    sum_of_diagonals_x = TR_BL_diagonal_x + TL_BR_diagonal_x
+    sum_of_diagonals_y = TR_BL_diagonal_y + TL_BR_diagonal_y
+
+    # Angle of sum-of-diagonals vector with the North-South direction
+    # Note: the minus sign changes convention from clockwise to counter-clockwise
+    angle = -np.arctan2(sum_of_diagonals_x, sum_of_diagonals_y)  # = - atan(x/y)
+
+    # Convert to degrees and assign to angles_arr
+    angles_arr = np.rad2deg(angle)
+
+    # Assign angles_arr to supergrid
+    t_angles = xr.DataArray(
+        angles_arr,
+        dims=["nyp", "nxp"],
+        coords={
+            "nyp": point.nyp.values,
+            "nxp": point.nxp.values,
+        },
+    )
+    return t_angles
+
+
+def modulo_around_point(x, x0, L):
+    """
+    Returns the modulo-:math:`L` value of :math:`x` within the interval :math:`[x_0 - L/2, x_0 + L/2]`.
+    If :math:`L ≤ 0`, then method returns :math:`x`.
+
+    (Adapted from MOM6 code; https://github.com/mom-ocean/MOM6/blob/776be843e904d85c7035ffa00233b962a03bfbb4/src/initialization/MOM_shared_initialization.F90#L592-L606)
+
+    Parameters
+    ----------
+    x: xr.DataArray
+       Value(s) to which to apply modulo arithmetic
+    x0: xr.DataArray
+        Center(s) of modulo range
+    L: float
+       Modulo range width
+
+    Returns
+    -------
+    float
+        ``x`` shifted by an integer multiple of ``L`` to be closer to ``x0``, i.e., within the interval ``[x0 - L/2, x0 + L/2]``
+    """
+    if L <= 0:
+        return x
+    else:
+        # Find that boundary point x0 + L/2
+        edge_indexes = np.where((x == x0 + L / 2))
+
+        # Modulo calculation
+        calc = ((x - (x0 - L / 2)) % L) + (x0 - L / 2)
+
+        # Find that boundary point x0 + L/2 does not flip to x0 - L/2
+        calc[edge_indexes] = x[edge_indexes]
+
+        return calc
